# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Golyv Fleet Booking API
  version: 1.0.0
  description: RESTful API for segment-aware bus seat booking across Egypt routes.
servers:
  - url: http://localhost:8000/api
paths:
...
  /trips/{trip}/available-seats:
    get:
      summary: Get available seats for a trip segment
      operationId: getAvailableSeats
      parameters:
        - name: trip
          in: path
          required: true
          schema:
            type: integer
        - name: start_station_id
          in: query
          required: true
          schema:
            type: integer
        - name: end_station_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seat availability for the segment
        "422":
          description: Invalid trip segment
        "404":
          description: Trip not found
...